requirements.txt
google-generativeai
pinecone
sentence-transformers
python-dotenv
google-adk
python-docx
fastapi
uvicorn
